Updated document management

+++ src/Controller/User/DocumentController.php
@@ -3,17 +3,20 @@
 namespace App\Controller\User;
 
 use App\Controller\Traits\Authorization;
+use App\Entity\Annotation;
 use App\Entity\Document;
 use App\Entity\Project;
 use App\Entity\User;
 use App\Form\DocumentType;
 use App\Form\ImportDocumentType;
+use App\Repository\AnnotationRepository;
 use App\Repository\DocumentRepository;
 use App\Repository\TagRepository;
 use App\Service\DocumentService;
 use App\Service\TextProcessor;
 use Doctrine\DBAL\Connection;
 use Doctrine\ORM\EntityManagerInterface;
+use Knp\Component\Pager\PaginatorInterface;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -40,8 +43,10 @@ class DocumentController extends AbstractController
         Document $document,
         Request $request,
         DocumentService $documentService,
+        DocumentRepository $documentRepository,
         TextProcessor $textProcessor,
-        TagRepository $tagRepository): Response
+        TagRepository $tagRepository,
+        PaginatorInterface $paginator): Response
     {
         /** @var User $user */
         $user = $this->getUser();
@@ -60,14 +65,24 @@ class DocumentController extends AbstractController
             return $this->redirectToRoute('user_document', ['id' => $document->getId()]);
         }
 
+        $documents = $paginator->paginate(
+            $documentRepository->createQueryBuilder('d')
+                ->where('d.project = :project')
+                ->setParameter('project', $document->getProject()),
+            $request->query->get('page', 1),
+            25
+        );
+
         return $this->render('user/document/index.html.twig', [
             'document' => $document,
+            'documents' => $documents,
             'form' => $form->createView(),
             'projectRole' => $this->getRole($user, $document->getProject()),
             'canEdit' => $this->canEdit($user, $document->getProject()),
             'lexicon' => $textProcessor->countWords($document->getWords()),
             'links' => $documentService->getLinks($document, $request),
             'tagTree' => $tagRepository->getProjectTags($document->getProject(), true),
+            'search' => $request->query->get('q'),
         ]);
     }
 
@@ -81,7 +96,8 @@ class DocumentController extends AbstractController
         Project $project,
         Request $request,
         DocumentService $documentService,
-        TextProcessor $textProcessor): Response
+        TextProcessor $textProcessor,
+        TagRepository $tagRepository): Response
     {
         /** @var User $user */
         $user = $this->getUser();
@@ -109,6 +125,7 @@ class DocumentController extends AbstractController
             'projectRole' => $this->getRole($user, $document->getProject()),
             'canEdit' => $this->canEdit($user, $project),
             'lexicon' => $textProcessor->countWords($document->getWords()),
+            'tagTree' => $tagRepository->getProjectTags($document->getProject(), true),
         ]);
     }
 
@@ -154,7 +171,7 @@ class DocumentController extends AbstractController
 
     /**
      * @Route("/user/project/{id}/import",
-     *     name="user_project_import_documents",
+     *     name="user_import_documents",
      *     requirements={"id": "\d+"})
      */
     public function importDocuments(
@@ -197,7 +214,7 @@ class DocumentController extends AbstractController
 
     /**
      * @Route("/user/project/{id}/delete-documents",
-     *     name="user_project_delete_documents",
+     *     name="user_delete_documents",
      *     requirements={"id": "\d+"})
      */
     public function deleteDocuments(
@@ -225,28 +242,48 @@ class DocumentController extends AbstractController
     }
 
     /**
-     * @Route("/user/project/{projectId}/delete-document/{documentId}",
-     *     name="user_project_delete_document",
-     *     requirements={
-     *         "projectId": "\d+",
-     *         "documentId": "\d+"
-     *     })
-     * @ParamConverter("project", options={"mapping": {"projectId": "id"}})
-     * @ParamConverter("document", options={"mapping": {"documentId": "id"}})
+     * @Route("/user/delete-document/{id}", name="user_delete_document")
      */
     public function deleteDocument(
-        Project $project,
         Document $document,
         EntityManagerInterface $entityManager): RedirectResponse
     {
         /** @var User $user */
         $user = $this->getUser();
+        $project = $document->getProject();
 
-        if ($this->canEdit($user, $project)) {
+        if ($project && $this->canEdit($user, $project)) {
             $entityManager->remove($document);
             $entityManager->flush();
             $this->addFlash('info', sprintf("Le document %s a été supprimé", $document->getTitle()));
-            return $this->redirectToRoute('user_view_project', ['id' => $project->getId()]);
         }
+
+        $this->addFlash('danger', sprintf("Vous n'avez pas la permission de supprimer le document %s", $document->getTitle()));
+        return $this->redirectToRoute('user_projects');
+    }
+
+    /**
+     * @Route("/user/tag-document/{id}", name="user_tag_document")
+     */
+    public function tagDocument(
+        Document $document,
+        Request $request,
+        AnnotationRepository $annotationRepository,
+        TagRepository $tagRepository,
+        EntityManagerInterface $entityManager): RedirectResponse
+    {
+        if (($request->request->get('action') === 'tag-document') && ($tag = $tagRepository->find($request->request->get('tag')))) {
+            $annotation = new Annotation();
+            $annotation->setContent($request->request->get('selection'));
+            $annotation->setDocument($document);
+            $annotation->setTag($tag);
+            $entityManager->persist($annotation);
+            $entityManager->flush();
+            $this->addFlash('success', "L'annotation a été sauvegardée");
+        } else {
+            $this->addFlash('danger', "L'annotation n'a pas été sauvegardée");
+        }
+
+        return $this->redirectToRoute('user_document', ['id' => $document->getId()]);
     }
 }
